+++ example/a.ts
@@ -1,9 +1,13 @@
 import { Elysia, t } from '../src'
+import { req } from '../test/utils'
 
 const app = new Elysia()
-	.get('/', ({ request }) => {
-		request.url
+	.onRequest(() => 'a')
+	.get('/:hash', async ({ params: { hash }, error, set }) => {
+		const file = Bun.file('example/teapot.webp')
+
+		return file
 	})
 	.listen(3000)
 
-// console.log(app.routes[0].compile().toString())
+console.log(app.routes[0].compile().toString())